spectate_server/requirement.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.12.2
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
